# comparison for go1.21+ with new slices.BinarySearch