--- a/.github/workflows/bench.yml
+++ b/.github/workflows/bench.yml
@@ -20,5 +20,10 @@ jobs:
       - uses: actions/setup-go@v4
         with:
           go-version: ${{ matrix.go }}
+          cache: false
       - run: go install golang.org/x/perf/cmd/benchstat@latest
-      - run: go test -bench=. -count=10 -cpu=1 | benchstat -col="/lib@(std xsort)" -
+      - run: go test -bench=. -count=6 -cpu=1 | benchstat -col="/lib" - | tee bench_results.txt
+      - uses: actions/upload-artifact@v3
+        with:
+          name: benchresults-go${{ matrix.go }}
+          path: bench_results.txt