--- a/.github/workflows/bench.yml
+++ b/.github/workflows/bench.yml
@@ -13,7 +13,7 @@ jobs:
           - "1.18"
           - "1.19"
           - "1.20"
-          - "1.21.0-rc.3"
+          - "1.21"
     name: Go ${{ matrix.go }} bench
     steps:
       - uses: actions/checkout@v3