--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,3 @@
 module github.com/mroth/xsort
 
-go 1.13
+go 1.18