Go 1.21 will add a new slices package, which notably contains a generics implementation of [`slices.BinarySearch`](https://pkg.go.dev/slices#BinarySearch), which should hopefully have improved performance characteristics that could make this library unnecessary.