@mroth 2023-07-26
Results on local laptop (Apple M2 Pro, working in a Docker container):
```
$ go1.21rc3 test -bench=. -count=6 -cpu=1 | benchstat -col="/lib" -
goos: linux
goarch: arm64
pkg: github.com/mroth/xsort
               │     sort     │               slices                │               xsort                │
               │    sec/op    │    sec/op     vs base               │   sec/op     vs base               │
SearchInts        59.96n ± 0%    15.06n ± 2%  -74.88% (p=0.002 n=6)   13.88n ± 0%  -76.85% (p=0.002 n=6)
SearchFloat64s    84.19n ± 0%    16.60n ± 0%  -80.28% (p=0.002 n=6)   13.89n ± 0%  -83.50% (p=0.002 n=6)
SearchStrings    136.00n ± 0%   126.00n ± 0%   -7.35% (p=0.002 n=6)   69.75n ± 0%  -48.71% (p=0.002 n=6)
geomean           88.22n         31.58n       -64.20%                 23.78n       -73.04%
```

Results are very promising for `int` and `float64`, but **unfortunately binary search for strings appears to be almost twice the time here for `slices` vs `xsort`**.

Looking at the [artifacts from the CI run](https://github.com/mroth/xsort/actions/runs/5663250069), Intel CPU results are similar, with `slices` even _underperforming_ the old `sort` package a bit:
```
goos: linux
goarch: amd64
pkg: github.com/mroth/xsort
cpu: Intel(R) Xeon(R) CPU E5-2673 v4 @ 2.30GHz
               │    sort     │               slices                │               xsort                │
               │   sec/op    │    sec/op     vs base               │   sec/op     vs base               │
SearchInts       95.19n ± 3%   32.45n ± 13%  -65.91% (p=0.002 n=6)   28.05n ± 2%  -70.53% (p=0.002 n=6)
SearchFloat64s   90.50n ± 5%   49.82n ±  3%  -44.95% (p=0.002 n=6)   30.47n ± 3%  -66.33% (p=0.002 n=6)
SearchStrings    219.7n ± 5%   237.9n ±  3%   +8.28% (p=0.002 n=6)   176.0n ± 3%  -19.87% (p=0.002 n=6)
geomean          123.7n        72.73n        -41.21%                 53.19n       -57.00%
```

This was a bit perplexing to me at first, but `slices.BinarySearch()` also changes the function signature a bit, returning a secondary bool to indicate whether the exact target was found.  Looking at the source code, the rest of the function is nearly identical to the old sort wrappers, but the [final return line](https://cs.opensource.google/go/go/+/master:src/slices/sort.go;l=140;drc=862fa6d099fb046e90efd537b2c0ac2667c23d90) is interesting:
```go
return i, i < n && (x[i] == target || (isNaN(x[i]) && isNaN(target)))
```

I suspect the simple `x[i] == target` may be the culprit here, as it adds a potentially expensive string comparison -- but is a string comparison really that expensive compared to a binary search over 10M elements?