@mroth 2023-07-26
...answer: hmm.

```
$ go tool pprof cpu.prof
File: xsort.test
Type: cpu
Time: Jul 25, 2023 at 9:06pm (EDT)
Duration: 2.48s, Total samples = 2.33s (93.86%)
Entering interactive mode (type "help" for commands, "o" for options)

(pprof) top5
Showing nodes accounting for 1520ms, 65.24% of 2330ms total
Dropped 21 nodes (cum <= 11.65ms)
Showing top 5 nodes out of 42
      flat  flat%   sum%        cum   cum%
     740ms 31.76% 31.76%      740ms 31.76%  cmpbody
     260ms 11.16% 42.92%      260ms 11.16%  memeqbody
     200ms  8.58% 51.50%      220ms  9.44%  fmt.(*fmt).fmtInteger
     190ms  8.15% 59.66%     1340ms 57.51%  slices.BinarySearch[go.shape.[]string,go.shape.string]
     130ms  5.58% 65.24%      220ms  9.44%  runtime.greyobject

(pprof) list slices.BinarySearch
Total: 2.33s
ROUTINE ======================== slices.BinarySearch[go.shape.[]string,go.shape.string] in /home/vscode/sdk/go1.21rc3/src/slices/sort.go
     190ms      1.34s (flat, cum) 57.51% of Total
         .          .    122:func BinarySearch[S ~[]E, E cmp.Ordered](x S, target E) (int, bool) {
         .          .    123:   // Inlining is faster than calling BinarySearchFunc with a lambda.
         .          .    124:   n := len(x)
         .          .    125:   // Define x[-1] < target and x[n] >= target.
         .          .    126:   // Invariant: x[i-1] < target, x[j] >= target.
         .          .    127:   i, j := 0, n
     120ms      120ms    128:   for i < j {
         .          .    129:           h := int(uint(i+j) >> 1) // avoid overflow when computing h
         .          .    130:           // i ≤ h < j
      60ms      1.20s    131:           if cmp.Less(x[h], target) {
         .          .    132:                   i = h + 1 // preserves x[i-1] < target
         .          .    133:           } else {
         .          .    134:                   j = h // preserves x[j] >= target
         .          .    135:           }
         .          .    136:   }
         .          .    137:   // i == j, x[i-1] < target, and x[j] (= x[i]) >= target  =>  answer is i.
      10ms       20ms    138:   return i, i < n && (x[i] == target || (isNaN(x[i]) && isNaN(target)))
         .          .    139:}
         .          .    140:
         .          .    141:// BinarySearchFunc works like [BinarySearch], but uses a custom comparison
         .          .    142:// function. The slice must be sorted in increasing order, where "increasing"
         .          .    143:// is defined by cmp. cmp should return 0 if the slice element matches
```

Nothing clear yet.  The majority of time is spent in string comparisons in cmp.Less, which makes sense?