@mroth 2023-08-10
Updated comparisons with release version of `go1.21.0`:
```
$ benchstat -col="/lib" bench_go121.txt
goos: darwin
goarch: arm64
pkg: github.com/mroth/xsort
               │     sort     │               slices                │               xsort                │
               │    sec/op    │    sec/op     vs base               │   sec/op     vs base               │
SearchInts        60.20n ± 1%    14.99n ± 0%  -75.11% (p=0.002 n=6)   13.86n ± 1%  -76.98% (p=0.002 n=6)
SearchFloat64s    84.14n ± 0%    16.57n ± 0%  -80.31% (p=0.002 n=6)   13.82n ± 0%  -83.57% (p=0.002 n=6)
SearchStrings    137.50n ± 1%   126.25n ± 1%   -8.18% (p=0.002 n=6)   69.41n ± 1%  -49.52% (p=0.002 n=6)
geomean           88.64n         31.53n       -64.43%                 23.69n       -73.27%

$ benchstat -col="/lib@(slices xsort)" bench_go121.txt
goos: darwin
goarch: arm64
pkg: github.com/mroth/xsort
               │    slices    │               xsort                │
               │    sec/op    │   sec/op     vs base               │
SearchInts        14.99n ± 0%   13.86n ± 1%   -7.51% (p=0.002 n=6)
SearchFloat64s    16.57n ± 0%   13.82n ± 0%  -16.60% (p=0.002 n=6)
SearchStrings    126.25n ± 1%   69.41n ± 1%  -45.02% (p=0.002 n=6)
geomean           31.53n        23.69n       -24.87%
```